+++ k8s.tf
@@ -6,14 +6,14 @@
 
 locals {
   heapsize      = ceil((var.memory * 85) / 100)
-  memory_limit  = ceil((var.memory * 120) / 100)
+  memory_limit  = length(var.memory_limit) != 0 ? var.memory_limit : ceil((var.memory * 120) / 100)
   actuator_port = 18000
   wd_port       = 48869
   k8s_cpu       = var.cpu / 1024
-  k8s_cpu_limit = (var.cpu / 1024) * 1.25
+  k8s_cpu_limit = length(var.cpu_limit) != 0 ? var.cpu_limit / 1024 : (var.cpu / 1024) * 1.25
 }
 
-resource "kubernetes_service_account" "waggle_dance" {
+resource "kubernetes_service_account_v1" "waggle_dance" {
   count = var.wd_instance_type == "k8s" ? 1 : 0
   metadata {
     name        = local.instance_alias
@@ -22,7 +22,23 @@ resource "kubernetes_service_account" "waggle_dance" {
       "eks.amazonaws.com/role-arn" = var.oidc_provider == "" ? "" : aws_iam_role.waggle_dance_k8s_role_iam[0].arn
     }
   }
-  automount_service_account_token = true
+}
+
+resource "kubernetes_secret_v1" "waggle_dance" {
+  count = var.wd_instance_type == "k8s" ? 1 : 0
+  metadata {
+    name        = local.instance_alias
+    namespace   = var.k8s_namespace
+    annotations = {
+      "kubernetes.io/service-account.name"      = local.instance_alias
+      "kubernetes.io/service-account.namespace" = var.k8s_namespace
+    }
+  }
+  type = "kubernetes.io/service-account-token"
+
+  depends_on = [
+    kubernetes_service_account_v1.waggle_dance
+  ]
 }
 
 resource "kubernetes_deployment_v1" "waggle_dance" {
@@ -49,17 +65,18 @@ resource "kubernetes_deployment_v1" "waggle_dance" {
           name = local.instance_alias
         }
         annotations = {
-          "ad.datadoghq.com/waggledance.check_names" = var.datadog_metrics_enabled ?  "[\"openmetrics\"]" : null
-          "ad.datadoghq.com/waggledance.init_configs" = var.datadog_metrics_enabled ?  "[{}]" : null
-          "ad.datadoghq.com/waggledance.instances" = var.datadog_metrics_enabled ?  "[{ \"prometheus_url\": \"http://%%host%%:${var.metrics_port}/actuator/prometheus\", \"namespace\": \"waggledance\", \"metrics\": [ \"${join("\",\"", var.datadog_metrics_waggledance)}\" ]  }]" : null
+          "ad.datadoghq.com/${local.instance_alias}.check_names" = var.datadog_metrics_enabled ?  "[\"openmetrics\"]" : null
+          "ad.datadoghq.com/${local.instance_alias}.init_configs" = var.datadog_metrics_enabled ?  "[{}]" : null
+          "ad.datadoghq.com/${local.instance_alias}.instances" = var.datadog_metrics_enabled ?  "[{ \"prometheus_url\": \"http://%%host%%:${var.metrics_port}/actuator/prometheus\", \"namespace\": \"waggledance\", \"metrics\": [ \"${join("\",\"", var.datadog_metrics_waggledance)}\" ]  }]" : null
           "prometheus.io/scrape" : var.prometheus_enabled
           "prometheus.io/port" : local.actuator_port
           "prometheus.io/path" : "/actuator/prometheus"
+          "iam.amazonaws.com/role" = var.oidc_provider == "" ? aws_iam_role.waggle_dance_k8s_role_iam[0].name : null
         }
       }
 
       spec {
-        service_account_name            = kubernetes_service_account.waggle_dance[0].metadata.0.name
+        service_account_name            = kubernetes_service_account_v1.waggle_dance[0].metadata.0.name
         automount_service_account_token = true
         dynamic "security_context"  {
           for_each = var.enable_sysctl_config_in_eks ? ["enabled"] : []