+++ CHANGELOG.md
@@ -3,9 +3,38 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [4.2.3] - 2024-01-19
+
+## [4.5.3] - 2024-06-28
 ### Fixed
 - Enable optional sysctl settings for EKS. To overwrite TCP keepalive settings.
+
+## [4.5.2] - 2024-06-04
+### Updated
+- Changed Service account creation to make it work with eks 1.24 and later.
+
+## [4.5.1] - 2024-05-08
+### Added
+- Adding tags to the Datadog agent
+
+## [4.5.0] - 2024-04-25
+### Added
+- Added optional `extended_server_config`. For instance to configure Waggle Dance Rate limiting configuration.
+
+## [4.4.0] - 2024-04-18
+### Added
+- Add new variable `cpu_limit` and `memory_limit` to support set pod resource guaranteed.
+    
+## [4.3.2] - 2024-03-26
+### Updated
+- Removed the datadog explicit provider configuration as suggested here: https://developer.hashicorp.com/terraform/language/modules/develop/providers#legacy-shared-modules-with-provider-configurations
+
+## [4.3.1] - 2024-02-22
+### Updated
+- Updated eks to use instance name instead of hard coding.
+
+## [4.3.0] - 2024-02-08
+### Added
+- Variable to add a readonly primary optional host/port. Waggle Dance will redirect get calls automatically to those.
 
 ## [4.2.2] - 2024-01-05
 ### Fixed